apps/traffic-generator/requirements.txt
requests>=2.31.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
opentelemetry-instrumentation-requests>=0.45b0
